// packages/test-force_optimize_speed-long_type_string/gen/google/type/datetime.ts
export interface DateTime {
    /**
     * Optional. Year of date. Must be from 1 to 9999, or 0 if specifying a
     * datetime without a year.
     *
     * @generated from protobuf field: int32 year = 1;
     */
    year: number;
    /**
     * Required. Month of year. Must be from 1 to 12.
     *
     * @generated from protobuf field: int32 month = 2;
     */
    month: number;
    /**
     * Required. Day of month. Must be from 1 to 31 and valid for the year and
     * month.
     *
     * @generated from protobuf field: int32 day = 3;
     */
    day: number;
    /**
     * Required. Hours of day in 24 hour format. Should be from 0 to 23. An API
     * may choose to allow the value "24:00:00" for scenarios like business
     * closing time.
     *
     * @generated from protobuf field: int32 hours = 4;
     */
    hours: number;
    /**
     * Required. Minutes of hour of day. Must be from 0 to 59.
     *
     * @generated from protobuf field: int32 minutes = 5;
     */
    minutes: number;
    /**
     * Required. Seconds of minutes of the time. Must normally be from 0 to 59. An
     * API may allow the value 60 if it allows leap-seconds.
     *
     * @generated from protobuf field: int32 seconds = 6;
     */
    seconds: number;
    /**
     * Required. Fractions of seconds in nanoseconds. Must be from 0 to
     * 999,999,999.
     *
     * @generated from protobuf field: int32 nanos = 7;
     */
    nanos: number;
    /**
     * @generated from protobuf oneof: time_offset
     */
    timeOffset: {
        oneofKind: "utcOffset";
        /**
         * UTC offset. Must be whole seconds, between -18 hours and +18 hours.
         * For example, a UTC offset of -4:00 would be represented as
         * { seconds: -14400 }.
         *
         * @generated from protobuf field: google.protobuf.Duration utc_offset = 8;
         */
        utcOffset: Duration;
    } | {
        oneofKind: "timeZone";
        /**
         * Time zone.
         *
         * @generated from protobuf field: google.type.TimeZone time_zone = 9;
         */
        timeZone: TimeZone;
    } | {
        oneofKind: undefined;
    };
}
/**
 * Represents a time zone from the
 * [IANA Time Zone Database](https://www.iana.org/time-zones).
 *
 * @generated from protobuf message google.type.TimeZone
 */
export interface TimeZone {
    /**
     * IANA Time Zone Database time zone, e.g. "America/New_York".
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * Optional. IANA Time Zone Database version number, e.g. "2019a".
     *
     * @generated from protobuf field: string version = 2;
     */
    version: string;
}
// ...
export const DateTime = new DateTime$Type();
// @generated message type with reflection information, may provide speed optimized methods
class TimeZone$Type extends MessageType<TimeZone> {
    constructor() {
        super("google.type.TimeZone", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "version", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<TimeZone>): TimeZone {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = "";
        message.version = "";
        if (value !== undefined)
            reflectionMergePartial<TimeZone>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: TimeZone): TimeZone {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string id */ 1:
                    message.id = reader.string();
                    break;
                case /* string version */ 2:
                    message.version = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: TimeZone, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string id = 1; */
        if (message.id !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.id);
        /* string version = 2; */
        if (message.version !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.version);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message google.type.TimeZone
 */
export const TimeZone = new TimeZone$Type();